updating reservation

+++ P2/restify/properties/views/deletereservation.py
@@ -1,4 +1,5 @@
 from django.shortcuts import get_object_or_404
+from accounts.models import CustomUser
 from properties.models import Reservation
 from properties.serializers import ReservationSerializer
 from rest_framework.response import Response
@@ -10,7 +11,13 @@ class DeleteReservation(DestroyAPIView):
     serializer = ReservationSerializer
     permission_classes = [IsAuthenticated]
 
-    def delete(self):
+    def delete(self, *args, **kwargs):
         reservation = get_object_or_404(Reservation, id=self.kwargs['pk'])
-        reservation.delete()
-        return Response(status=200)
+        user = get_object_or_404(CustomUser, pk=self.request.user.id)
+        if user == reservation.property.owner:
+            reservation.delete()
+            return Response(status=200)
+        else:
+            return Response([{
+                'details': 'Permission denied'
+            }])
